placeholders removed and inputContainer moved to parent folder
