+++ src/components/sidebar/inputContainer.scss
@@ -1,0 +1,2 @@
+@import '../shared/styles/common';
+@import '../shared/styles/form';
